phone update
In Viewcontroller.m there is a function that would use the phone there
are two things that need to be done: one. the phone number has to be
called from the database and match the secOff value which has to match
school and user location
two. we got to be able to run from background and make a widget so that
it can be easily accessed the voice assistant functionality will not be
available yet